auth-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
passlib
bcrypt==4.0.1
PyJWT
python-multipart